server/requirements.txt
# ChannelZero API — Python dependencies
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
asyncpg>=0.30,<1.0
pydantic[email]>=2.7,<3.0
pydantic-settings>=2.3,<3.0
python-dotenv>=1.0,<2.0
argon2-cffi>=23.1,<24.0
PyJWT>=2.8,<3.0
cryptography>=43.0,<44.0
httpx>=0.27,<1.0
